UPDATE move constant file to avoid cyclic dependecies
