+++ google-datacatalog-rdbms-connector/src/google/datacatalog_connectors/rdbms/prepare/sql_objects/sql_objects_datacatalog_tag_factory.py
@@ -18,7 +18,7 @@ from google.cloud import datacatalog
 
 from google.datacatalog_connectors.commons import prepare
 
-from google.datacatalog_connectors.rdbms.scrape import constants
+from google.datacatalog_connectors.rdbms.common import constants
 
 
 class SQLObjectsDataCatalogTagFactory(prepare.BaseTagFactory):
