+++ google-datacatalog-rdbms-connector/src/google/datacatalog_connectors/rdbms/scrape/query_assembler.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from google.datacatalog_connectors.rdbms.scrape import constants
+from google.datacatalog_connectors.rdbms.common import constants
 
 
 class QueryAssembler:
