+++ google-datacatalog-rdbms-connector/tests/google/datacatalog_connectors/rdbms/test_utils/test_utils.py
@@ -16,10 +16,13 @@
 
 import os
 
+
 from google.datacatalog_connectors.commons_test import utils
 from google.datacatalog_connectors.rdbms import datacatalog_cli
+from google.datacatalog_connectors.rdbms.common import constants
 from google.datacatalog_connectors.rdbms.scrape import \
-     constants, metadata_enricher, metadata_scraper, query_assembler
+     metadata_enricher, metadata_scraper, query_assembler
+
 import mock
